refactor: factorized structs between compiled.SparseR1Cs and compiled.R1CS
