+++ internal/backend/compiled/r1c.go
@@ -16,11 +16,13 @@ package compiled
 
 import (
 	"math/big"
-	"strconv"
 	"strings"
-
-	"github.com/consensys/gnark/backend/hint"
 )
+
+// R1C used to compute the wires
+type R1C struct {
+	L, R, O LinearExpression
+}
 
 // LinearExpression represent a linear expression of variables
 type LinearExpression []Term
@@ -52,32 +54,6 @@ func (l LinearExpression) Less(i, j int) bool {
 	return iVis > jVis
 }
 
-// R1C used to compute the wires
-type R1C struct {
-	L, R, O LinearExpression
-}
-
-// Visibility encodes a Variable (or wire) visibility
-// Possible values are Unset, Internal, Secret or Public
-type Visibility uint8
-
-const (
-	Unset Visibility = iota
-	Internal
-	Secret
-	Public
-	Virtual
-)
-
-// Hint represents a solver hint
-// it enables the solver to compute a Wire with a function provided at solving time
-// using pre-defined inputs
-type Hint struct {
-	WireID int                // resulting wire ID to compute
-	ID     hint.ID            // hint function id
-	Inputs []LinearExpression // terms to inject in the hint function
-}
-
 func (r1c *R1C) String(coeffs []big.Int) string {
 	var sbb strings.Builder
 	sbb.WriteString("L[")
@@ -98,24 +74,4 @@ func (l LinearExpression) string(sbb *strings.Builder, coeffs []big.Int) {
 			sbb.WriteString(" + ")
 		}
 	}
-}
-
-func (t Term) string(sbb *strings.Builder, coeffs []big.Int) {
-	sbb.WriteString(coeffs[t.CoeffID()].String())
-	sbb.WriteString("*")
-	switch t.VariableVisibility() {
-	case Internal:
-		sbb.WriteString("i")
-	case Public:
-		sbb.WriteString("p")
-	case Secret:
-		sbb.WriteString("s")
-	case Virtual:
-		sbb.WriteString("v")
-	case Unset:
-		sbb.WriteString("u")
-	default:
-		panic("not implemented")
-	}
-	sbb.WriteString(strconv.Itoa(t.VariableID()))
 }
