+++ internal/backend/compiled/r1cs_sparse.go
@@ -14,86 +14,13 @@
 
 package compiled
 
-import (
-	"io"
-
-	"github.com/consensys/gnark-crypto/ecc"
-)
-
-// SparseR1CS represents a Plonk like circuit
+// R1CS decsribes a set of SparseR1C constraint
 type SparseR1CS struct {
-
-	// Variables [publicVariables| secretVariables | internalVariables ]
-	NbInternalVariables int
-	NbPublicVariables   int
-	NbSecretVariables   int
-
-	// Constraints contains an ordered list of SparseR1C
-	// the solver will iterate through them and is guaranteed that there will be at most one
-	// unsolved wire per constraint
+	CS
 	Constraints []SparseR1C
-
-	// Logs (e.g. variables that have been printed using cs.Println)
-	Logs      []LogEntry
-	DebugInfo []LogEntry
-
-	// Hints
-	Hints []Hint
-
-	MHints map[int]int // maps wire id to hint id
-	MDebug map[int]int // maps constraint id to debugInfo id
-}
-
-// GetNbVariables return number of internal, secret and public variables
-func (cs *SparseR1CS) GetNbVariables() (internal, secret, public int) {
-	internal = cs.NbInternalVariables
-	secret = cs.NbSecretVariables
-	public = cs.NbPublicVariables
-	return
 }
 
 // GetNbConstraints returns the number of constraints
 func (cs *SparseR1CS) GetNbConstraints() int {
 	return len(cs.Constraints)
-}
-
-// GetNbWires returns the number of wires (internal)
-func (cs *SparseR1CS) GetNbWires() int {
-	return cs.NbInternalVariables
-}
-
-// FrSize panics
-func (cs *SparseR1CS) FrSize() int {
-	panic("not implemented")
-}
-
-// GetNbCoefficients return the number of unique coefficients needed in the R1CS
-func (cs *SparseR1CS) GetNbCoefficients() int {
-	panic("not implemented")
-}
-
-// CurveID returns ecc.UNKNOWN as this is a untyped R1CS using big.Int
-func (cs *SparseR1CS) CurveID() ecc.ID {
-	return ecc.UNKNOWN
-}
-
-// WriteTo panics
-func (cs *SparseR1CS) WriteTo(w io.Writer) (n int64, err error) {
-	panic("not implemented")
-}
-
-// ReadFrom panics
-func (cs *SparseR1CS) ReadFrom(r io.Reader) (n int64, err error) {
-	panic("not implemented")
-}
-
-// SetLoggerOutput replace existing logger output with provided one
-// default uses os.Stdout
-// if nil is provided, logs are not printed
-func (cs *SparseR1CS) SetLoggerOutput(w io.Writer) {
-	panic("not implemented")
-}
-
-func (cs *SparseR1CS) ToHTML(w io.Writer) error {
-	panic("not implemtened")
 }
